scripts/requirements.txt
Tabulate
pyyaml